+++ packages/inngest/src/components/InngestStepTools.ts
@@ -30,13 +30,37 @@ import {
 } from "./Inngest.js";
 import { InngestFunction } from "./InngestFunction.js";
 import { InngestFunctionReference } from "./InngestFunctionReference.js";
+import { gemini, openai, type AiAdapter } from "./ai/index.js";
+
 import { type InngestExecution } from "./execution/InngestExecution.js";
 
 export interface FoundStep extends HashedOp {
   hashedId: string;
   fn?: (...args: unknown[]) => unknown;
   rawArgs: unknown[];
+
+  /**
+   * A boolean representing whether the step has been fulfilled, either
+   * resolving or rejecting the `Promise` returned to userland code.
+   *
+   * Note that this is distinct from {@link hasStepState}, which instead tracks
+   * whether the step has been given some state from the Executor. State from
+   * the Executor could be data other than a resolution or rejection, such as
+   * inputs.
+   */
   fulfilled: boolean;
+
+  /**
+   * A boolean representing whether the step has been given some state from the
+   * Executor. State from the Executor could be data other than a resolution or
+   * rejection, such as inputs.
+   *
+   * This is distinct from {@link fulfilled}, which instead tracks whether the
+   * step has been fulfilled, either resolving or rejecting the `Promise`
+   * returned to userland code.
+   */
+  hasStepState: boolean;
+
   handled: boolean;
 
   /**
@@ -49,6 +73,10 @@ export interface FoundStep extends HashedOp {
    * invocation.
    */
   handle: () => Promise<boolean>;
+
+  // TODO This is used to track the input we want for this step. Might be
+  // present in ctx from Executor.
+  input?: unknown;
 }
 
 export type MatchOpFn<
@@ -144,6 +172,73 @@ export const createStepTools = <TClient extends Inngest.Any>(
       const parsedArgs = args as unknown as [StepOptionsOrId, ...unknown[]];
       return stepHandler({ args: parsedArgs, matchOp, opts });
     }) as T;
+  };
+
+  /**
+   * Create a new step run tool that can be used to run a step function using
+   * `step.run()` as a shim.
+   */
+  const createStepRun = (
+    /**
+     * The sub-type of this step tool, exposed via `opts.type` when the op is
+     * reported.
+     */
+    type?: string
+  ) => {
+    return createTool<
+      <TFn extends (...args: Parameters<TFn>) => unknown>(
+        idOrOptions: StepOptionsOrId,
+
+        /**
+         * The function to run when this step is executed. Can be synchronous or
+         * asynchronous.
+         *
+         * The return value of this function will be the return value of this
+         * call to `run`, meaning you can return and reason about return data
+         * for next steps.
+         */
+        fn: TFn,
+
+        /**
+         * Optional input to pass to the function. If this is specified, Inngest
+         * will keep track of the input for this step and be able to display it
+         * in the UI.
+         */
+        ...input: Parameters<TFn>
+      ) => Promise<
+        /**
+         * TODO Middleware can affect this. If run input middleware has returned
+         * new step data, do not Jsonify.
+         */
+        SimplifyDeep<
+          Jsonify<
+            TFn extends (...args: Parameters<TFn>) => Promise<infer U>
+              ? Awaited<U extends void ? null : U>
+              : ReturnType<TFn> extends void
+                ? null
+                : ReturnType<TFn>
+          >
+        >
+      >
+    >(
+      ({ id, name }, _fn, ...input) => {
+        const opts: HashedOp["opts"] = {
+          ...(input.length ? { input } : {}),
+          ...(type ? { type } : {}),
+        };
+
+        return {
+          id,
+          op: StepOpCode.StepPlanned,
+          name: id,
+          displayName: name ?? id,
+          ...(Object.keys(opts).length ? { opts } : {}),
+        };
+      },
+      {
+        fn: (_, fn, ...input) => fn(...input),
+      }
+    );
   };
 
   /**
@@ -264,45 +359,74 @@ export const createStepTools = <TClient extends Inngest.Any>(
      * of the `run` tool, meaning you can return and reason about return data
      * for next steps.
      */
-    run: createTool<
-      <T extends () => unknown>(
-        idOrOptions: StepOptionsOrId,
+    run: createStepRun(),
 
-        /**
-         * The function to run when this step is executed. Can be synchronous or
-         * asynchronous.
-         *
-         * The return value of this function will be the return value of this
-         * call to `run`, meaning you can return and reason about return data
-         * for next steps.
-         */
-        fn: T
-      ) => Promise<
-        /**
-         * TODO Middleware can affect this. If run input middleware has returned
-         * new step data, do not Jsonify.
-         */
-        SimplifyDeep<
-          Jsonify<
-            T extends () => Promise<infer U>
-              ? Awaited<U extends void ? null : U>
-              : ReturnType<T> extends void
-                ? null
-                : ReturnType<T>
-          >
-        >
-      >
-    >(
-      ({ id, name }) => {
+    /**
+     * AI tooling for running AI models and other AI-related tasks.
+     */
+    ai: {
+      /**
+       * Use this tool to have Inngest make your AI calls. Useful for agentic workflows.
+       *
+       * Input is also tracked for this tool, meaning you can pass input to the
+       * function and it will be displayed and editable in the UI.
+       */
+      infer: createTool<
+        <TAdapter extends AiAdapter>(
+          idOrOptions: StepOptionsOrId,
+          options: AiInferOpts<TAdapter>
+        ) => Promise<AiAdapter.Output<TAdapter>>
+      >(({ id, name }, options) => {
+        const modelCopy = { ...options.model };
+
+        // Allow the model to mutate options and body for this call
+        options.model.onCall?.(modelCopy, options.body);
+
         return {
           id,
-          op: StepOpCode.StepPlanned,
-          name: id,
+          op: StepOpCode.AiGateway,
           displayName: name ?? id,
+          opts: {
+            type: "step.ai.infer",
+            url: modelCopy.url,
+            headers: modelCopy.headers,
+            auth_key: modelCopy.authKey,
+            format: modelCopy.format,
+            body: options.body,
+          },
         };
+      }),
+
+      /**
+       * Use this tool to wrap AI models and other AI-related tasks. Each call
+       * to `wrap` will be retried individually, meaning you can compose complex
+       * workflows that safely retry dependent asynchronous actions.
+       *
+       * Input is also tracked for this tool, meaning you can pass input to the
+       * function and it will be displayed and editable in the UI.
+       */
+      wrap: createStepRun("step.ai.wrap"),
+
+      /**
+       * Models for AI inference and other AI-related tasks.
+       */
+      models: {
+        /**
+         * Create an OpenAI model using the OpenAI chat format.
+         *
+         * By default it targets the `https://api.openai.com/v1/` base URL.
+         */
+        openai,
+
+        /**
+         * Create a Gemini model using the OpenAI chat format.
+         *
+         * By default it targets the `https://generativelanguage.googleapis.com/v1beta/`
+         * base URL.
+         */
+        gemini,
       },
-      { fn: (stepOptions, fn) => fn() }
-    ),
+    },
 
     /**
      * Wait a specified amount of time before continuing.
@@ -567,3 +691,33 @@ type WaitForEventOpts<
   "match",
   "if"
 >;
+
+/**
+ * Options for `step.ai.infer()`.
+ */
+type AiInferOpts<TModel extends AiAdapter> = {
+  /**
+   * The model to use for the inference. Create a model by importing from
+   * `"inngest"` or by using `step.ai.models.*`.
+   *
+   * @example Import `openai()`
+   * ```ts
+   * import { openai } from "inngest";
+   *
+   * const model = openai({ model: "gpt-4" });
+   * ```
+   *
+   * @example Use a model from `step.ai.models`
+   * ```ts
+   * async ({ step }) => {
+   *            const model = step.ai.models.openai({ model: "gpt-4" });
+   * }
+   * ```
+   */
+  model: TModel;
+
+  /**
+   * The input to pass to the model.
+   */
+  body: AiAdapter.Input<TModel>;
+};