+++ packages/inngest/src/components/execution/v1.ts
@@ -272,15 +272,18 @@ class V1InngestExecution extends InngestExecution implements IInngestExecution {
 
     if (
       op &&
-      op.op === StepOpCode.StepPlanned &&
-      typeof op.opts === "undefined"
+      op.op === StepOpCode.StepPlanned
+      // TODO We must individually check properties here that we do not want to
+      // execute on, such as retry counts. Nothing exists here that falls in to
+      // this case, but should be accounted for when we add them.
+      // && typeof op.opts === "undefined"
     ) {
       return op.hashedId;
     }
   }
 
   private async filterNewSteps(
-    steps: FoundStep[]
+    foundSteps: FoundStep[]
   ): Promise<[OutgoingOp, ...OutgoingOp[]] | void> {
     if (this.options.requestedRunStep) {
       return;
@@ -289,7 +292,7 @@ class V1InngestExecution extends InngestExecution implements IInngestExecution {
     /**
      * Gather any steps that aren't memoized and report them.
      */
-    const newSteps = steps.filter((step) => !step.fulfilled);
+    const newSteps = foundSteps.filter((step) => !step.fulfilled);
 
     if (!newSteps.length) {
       return;
@@ -300,8 +303,8 @@ class V1InngestExecution extends InngestExecution implements IInngestExecution {
      * steps. This may indicate that step presence isn't determinate.
      */
     const stepsToFulfil = Object.keys(this.state.stepState).length;
-    const fulfilledSteps = steps.filter((step) => step.fulfilled).length;
-    const foundAllCompletedSteps = stepsToFulfil === fulfilledSteps;
+    const knownSteps = foundSteps.filter((step) => step.hasStepState).length;
+    const foundAllCompletedSteps = stepsToFulfil === knownSteps;
 
     if (!foundAllCompletedSteps) {
       // TODO Tag
@@ -854,18 +857,58 @@ class V1InngestExecution extends InngestExecution implements IInngestExecution {
       const { promise, resolve, reject } = createDeferredPromise();
       const hashedId = _internals.hashId(opId.id);
       const stepState = this.state.stepState[hashedId];
+      let isFulfilled = false;
       if (stepState) {
         stepState.seen = true;
+
+        if (typeof stepState.input === "undefined") {
+          isFulfilled = true;
+        }
+      }
+
+      let extraOpts: Record<string, unknown> | undefined;
+      let fnArgs = [...args];
+
+      if (
+        typeof stepState?.input !== "undefined" &&
+        Array.isArray(stepState.input)
+      ) {
+        switch (opId.op) {
+          // `step.run()` has its function input affected
+          case StepOpCode.StepPlanned: {
+            // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
+            fnArgs = [...args.slice(0, 2), ...stepState.input];
+            // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
+            extraOpts = { input: [...stepState.input] };
+            break;
+          }
+
+          // `step.ai.infer()` has its body affected
+          case StepOpCode.AiGateway: {
+            extraOpts = {
+              body: {
+                ...(typeof opId.opts?.body === "object"
+                  ? { ...opId.opts.body }
+                  : {}),
+                ...stepState.input[0],
+              },
+            };
+            break;
+          }
+        }
       }
 
       const step: FoundStep = {
         ...opId,
-        rawArgs: args,
+        opts: { ...opId.opts, ...extraOpts },
+        rawArgs: fnArgs, // TODO What is the right value here? Should this be raw args without affected input?
         hashedId,
+        input: stepState?.input,
         // eslint-disable-next-line @typescript-eslint/no-unsafe-argument
-        fn: opts?.fn ? () => opts.fn?.(...args) : undefined,
+        fn: opts?.fn ? () => opts.fn?.(...fnArgs) : undefined,
         promise,
-        fulfilled: Boolean(stepState),
+        fulfilled: isFulfilled,
+        hasStepState: Boolean(stepState),
         displayName: opId.displayName ?? opId.id,
         handled: false,
         handle: async () => {
@@ -875,15 +918,16 @@ class V1InngestExecution extends InngestExecution implements IInngestExecution {
 
           step.handled = true;
 
-          if (stepState) {
+          if (isFulfilled && stepState) {
             stepState.fulfilled = true;
 
-            // For some execution scenarios such as testing, `data` and/or
-            // `error` may be `Promises`. This could also be the case for future
-            // middleware applications. For this reason, we'll make sure the
-            // values are fully resolved before continuing.
+            // For some execution scenarios such as testing, `data`, `error`,
+            // and `input` may be `Promises`. This could also be the case for
+            // future middleware applications. For this reason, we'll make sure
+            // the values are fully resolved before continuing.
             await stepState.data;
             await stepState.error;
+            await stepState.input;
 
             if (typeof stepState.data !== "undefined") {
               resolve(stepState.data);